Move each node to its own file
