+++ .gitignore
@@ -1,1 +1,2 @@
 custom-lists
+__pycache__
